add more functionality for logged in users, sort by marketcap, sort by volume traded & list top 10 crypto exchanges

+++ models/Cryptocurrency.js
@@ -17,7 +17,7 @@ Cryptocurrency.init(
       type: DataTypes.STRING,
     },
     price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(8,6),
       allowNull: false,
     },
     market_cap: {
